requirements.txt
streamlit
torch
transformers
pandas
Pillow
sentence-transformers
langchain
langchain-community
langchain-huggingface
beautifulsoup4
faiss-cpu
huggingface_hub
unstructured